Correct argument types in UIScriptController::activateDataListSuggestion after r252062 https://bugs.webkit.org/show_bug.cgi?id=203116
Reviewed by Tim Horton.

Use `unsigned` to represent `unsigned long` arguments declared in the IDL.

* DumpRenderTree/mac/UIScriptControllerMac.h:
* DumpRenderTree/mac/UIScriptControllerMac.mm:
(WTR::UIScriptControllerMac::activateDataListSuggestion):
* TestRunnerShared/UIScriptContext/UIScriptController.h:
(WTR::UIScriptController::activateDataListSuggestion):
* WebKitTestRunner/ios/UIScriptControllerIOS.h:
* WebKitTestRunner/ios/UIScriptControllerIOS.mm:
(WTR::UIScriptControllerIOS::activateDataListSuggestion):
* WebKitTestRunner/mac/UIScriptControllerMac.h:
* WebKitTestRunner/mac/UIScriptControllerMac.mm:
(WTR::UIScriptControllerMac::activateDataListSuggestion):


Canonical link: https://commits.webkit.org/217700@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252719 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/DumpRenderTree/mac/UIScriptControllerMac.h
@@ -50,7 +50,7 @@ public:
     void simulateAccessibilitySettingsChangeNotification(JSValueRef) override;
     NSUndoManager *platformUndoManager() const override;
     void copyText(JSStringRef) override;
-    void activateDataListSuggestion(long, JSValueRef) override;
+    void activateDataListSuggestion(unsigned, JSValueRef) override;
 };
 
 }
